requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
threadpoolctl==2.2.0
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
tqdm==4.62.2
typing-extensions==3.10.0.2
